List available examples on the landing page
