+++ src/pages/index.tsx
@@ -1,43 +1,27 @@
-import { NextPage, GetStaticProps } from 'next'
+import Link from 'next/link'
 import React from 'react'
-import StemAndLeafPlot from 'components/_shared/StemAndLeafPlot/StemAndLeafPlot'
-import Header from 'components/_shared/Header/Header'
 
-interface Props {
-  items: number[]
-}
-const Index: NextPage<Props> = (props) => {
+function Index() {
+  const PAGES = [
+    { url: 'graphic', title: 'Example of a complete graphic (Header+Chart)' },
+    { url: 'chart', title: 'Minimal example of a responsive chart' },
+    { url: 'map', title: 'Map example' }
+  ]
+
   return (
     <>
-      <Header>
-        {{
-          title:
-            'Ein Aussagesatz oder notfalls eine Kurzbeschreibung, wenn das eine Live-Grafik ist',
-          subtitle: 'Angabe der Metrik und Erklärung der Grafik',
-        }}
-      </Header>
-      <StemAndLeafPlot items={props.items} numberOfSteps={30} maxValue={30} />
+      <b>Example pages:</b>
+      <ul>
+        {PAGES.map(({ url, title }) => (
+          <li>
+            <Link href={url}>
+              <a>{title}</a>
+            </Link>
+          </li>
+        ))}
+      </ul>
     </>
   )
-}
-
-export const getStaticProps: GetStaticProps = async (ctx) => {
-  // get static props is called at build / export time,
-  // so this could be a compute heavy function or a call to e.g. the google api
-
-  // prettier-ignore
-  const items = [
-      0, 0, 0, 2, 2, 2, 2, 3, 4, 5, 5, 5, 6, 7, 8, 9, 9, 10, 11, 12, 12, 12,
-      13, 13, 13, 14, 14, 15, 16, 16, 16, 17, 18, 18, 19, 19, 19,
-      20, 20, 20, 21, 21, 21, 21, 22, 22, 22, 23, 24, 25, 25, 26, 26, 26,
-      27, 27, 27, 28, 28, 28, 28, 29, 29, 29, 30, 30, 30, 30, 30, 30
-    ]
-
-  return {
-    props: {
-      items,
-    },
-  }
 }
 
 export default Index
